Moved multiple files change to trait and added docs
